+++ apps/web/eslint.config.js
@@ -1,14 +1,4 @@
 import { nextJsConfig } from "@repo/eslint-config/next-js";
 
 /** @type {import("eslint").Linter.Config[]} */
-export default [
-  ...nextJsConfig,
-  {
-    files: ["next.config.js"],
-    languageOptions: {
-      globals: {
-        process: "readonly",
-      },
-    },
-  },
-];
+export default nextJsConfig;